changed glXCopyContext mask back to GLuint

+++ src/mesa/drivers/x11/glxapi.c
@@ -1,4 +1,4 @@
-/* $Id: glxapi.c,v 1.13 2000/02/25 16:46:01 brianp Exp $ */
+/* $Id: glxapi.c,v 1.14 2000/02/25 17:04:42 brianp Exp $ */
 
 /*
  * Mesa 3-D graphics library
@@ -169,8 +169,7 @@ XVisualInfo *glXChooseVisual(Display *dpy, int screen, int *list)
 }
 
 
-void glXCopyContext(Display *dpy, GLXContext src, GLXContext dst,
-                    unsigned long mask)
+void glXCopyContext(Display *dpy, GLXContext src, GLXContext dst, GLuint mask)
 {
    struct _glxapi_table *t = get_dispatch(dpy);
    if (!t)
